improve project structure

+++ supervised/tree/ensemble/_adaboost.py
@@ -1,5 +1,5 @@
 import numpy as np
-from .._base import SupervisedModel
+from supervised._base import SupervisedModel
 
 
 # TODO
@@ -7,7 +7,7 @@ class Adaboost(SupervisedModel):
     def __init__(self):
         pass
 
-    def fit(self, x: np.ndarray, y: np.ndarray, **kwargs) -> np.float:
+    def fit(self, x: np.ndarray, y: np.ndarray, **kwargs) -> float:
         pass
 
     def predict(self, x: np.ndarray, **kwargs) -> np.ndarray:
